+++ CHANGELOG.md
@@ -2,8 +2,15 @@
 
 All notable changes to the "palantir-java-format-vscode" extension will be documented in this file.
 
-Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
+<!-- Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file. -->
 
-## [Unreleased]
+<!-- ## [Unreleased] -->
+
+## v0.0.9 - 2025-07-30
+### Added 
+
+- config item `enableDebugMode`: When enabled, use vscode popup window to show the executing formatting command for replaying the execution result
+
+## v0.0.8 - 2025-07-30
 
 - Initial release